Rename pkg/systema/grpctun to pkg/dnet, rename a few of the members
