+++ pkg/dnet/borrowed_net_pipe.go
@@ -5,7 +5,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package grpctun // MODIFIED
+package dnet // MODIFIED
 
 import (
 	"sync"
